refactor(Logs): convert Mode enum class to plain enum and reorganize header includes
- Change log Mode from enum class to plain enum : uint8_t for simpler bitwise operations
- Remove unnecessary static_cast<unsigned int> conversions in isUsingSyslog/StandardLog/WindowsEventLog methods
- Move logcolors.h and loglevels.h includes from logbase.h consumers to applog.h and rpclog.h
- logbase.h now only depends on logmodes.h, reducing include coupling
- Consumers that need colors/levels explicitly include them, improving header modularity

+++ Mantids30/Program_Logs/logbase.cpp
@@ -59,17 +59,17 @@ void LogBase::setDebug(bool value)
 
 bool LogBase::isUsingSyslog() const
 {
-    return (m_logMode & static_cast<unsigned int>(Mode::SYSLOG)) == static_cast<unsigned int>(Mode::SYSLOG);
+    return (m_logMode & Mode::SYSLOG) == Mode::SYSLOG;
 }
 
 bool LogBase::isUsingStandardLog() const
 {
-    return (m_logMode & static_cast<unsigned int>(Mode::STANDARD)) == static_cast<unsigned int>(Mode::STANDARD);
+    return (m_logMode & Mode::STANDARD) == Mode::STANDARD;
 }
 
 bool LogBase::isUsingWindowsEventLog() const
 {
-    return (m_logMode & static_cast<unsigned int>(Mode::WINEVENTS)) == static_cast<unsigned int>(Mode::WINEVENTS);
+    return (m_logMode & Mode::WINEVENTS) == Mode::WINEVENTS;
 }
 
 void LogBase::printDate(FILE *fp) const
